--CAS - PIFSC Oracle Developer Environment: Initial Working Version--
Forked the DSC PIFSC Oracle Developer Environment (ODE) repository to customize the development environment for the CAS database schema and APEX app as well as any dependencies

Updated deployment_scripts/prepare_docker_project.sh to retrieve the CAS project files from the git repository and place them into the corresponding locations to be included in the docker image when it's built.

Updated docker/src/run_db_app_deployment.sh to uncomment the code that waits for APEX to be installed since there is an APEX app for the CAS project.  Added code to create and deploy the CAS schema.

Updated README.md to specify the information for the customized CAS ODE